refactor: 拆分 agents 模块内联组件、类型和工具函数至独立文件
- 将 SkillCard、InstalledAssistantCard 抽离为独立组件
- 提取 AssistantListItem 类型至 types.ts
- 提取 normalizeAssistantVersion 工具至 utils.ts
- 对齐 agents 模块与项目组件/类型/工具分层规范
